requirements.txt
flet
pystray
Pillow
